cleanup for mr
