+++ .idea/sqldialects.xml
@@ -1,6 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project version="4">
   <component name="SqlDialectMappings">
+    <file url="file://$PROJECT_DIR$/newdb.py" dialect="GenericSQL" />
     <file url="file://$PROJECT_DIR$/sqdb.py" dialect="GenericSQL" />
   </component>
 </project>
