--- example.json
@@ -1,27 +1,0 @@
-{
-  "Article": {
-    "id": 1,
-    "created_at": 1668873815.1195192,
-    "fake_rating": 6.5,
-    "fake_rating_community": 5.43,
-    "clickbait_rating": 6.4,
-    "clickbait_rating_community": 4.8,
-    "published_at": 1668875608.3761518,
-    "title": "SomeTitle",
-    "description": "blablablablalballalala",
-    "img_url": "https://images.ctfassets.net/hrltx12pl8hq/7yQR5uJhwEkRfjwMFJ7bUK/dc52a0913e8ff8b5c276177890eb0129/offset_comp_772626-opt.jpg",
-    "author": "ktoś",
-    "url": "https://przegladsportowy.onet.pl/pilka-nozna/ceny-w-polskiej-restauracji-w-katarze-przebily-wszystko-to-nie-najwieksze-zaskoczenie/hyjyx5r"
-  }
-}
-
-
-    "review_data":
-    {
-
-      "other_img_urls": "https://webkit.org/demos/srcset/image-1x.png",
-      "hate_speech_detected": true
-    }
-
-  }
-}
