+++ frontend/package.json
@@ -14,11 +14,13 @@
     "@types/react": "18.0.25",
     "@types/react-dom": "18.0.9",
     "axios": "^1.1.3",
+    "create-next-app": "^13.0.4",
     "eslint": "8.28.0",
     "eslint-config-next": "13.0.4",
-    "next": "13.0.4",
-    "react": "18.2.0",
-    "react-dom": "18.2.0",
+    "next": "^13.0.4",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-router": "^6.4.3",
     "typescript": "4.9.3"
   },
   "devDependencies": {
