+++ scrap.py
@@ -1,86 +1,8 @@
-import io
 import os
-
-#STT
-# HWR
-import requests
-import webbrowser
-import hashlib
-import io
 import requests
 from bs4 import BeautifulSoup
-# import matplotlib.pyplot as plt
 import re
 
-
-
-def detect_document(path,typ,detale,jsonx):
-    """Detects document features in an image."""
-    #print("odczyt pliku")
-    with io.open(path, 'rb') as image_file:
-        content = image_file.read()
-
-    image = vision.types.Image(content=content)
-
-    response=[]
-
-    if str(typ)=="pismo_odręczne":
-        
-        response = client.document_text_detection(image=image)
-        all_text = ''
-
-        texts = response.text_annotations
-        licz=0
-        odp=""
-        for text in texts:
-            #print('\n"{}"'.format(text.description))
-            if licz==0:
-                #print(text.description)
-                odp=text.description
-                licz=licz+1
-        #print("")
-        
-        '''
-        for page in response.full_text_annotation.pages:
-            for block in page.blocks:
-                if detale:
-                    print('\nBlock confidence: {}\n'.format(block.confidence))
-
-                for paragraph in block.paragraphs:
-                    if detale:
-                        print('Paragraph confidence: {}'.format(paragraph.confidence))
-
-                    for word in paragraph.words:
-                        word_text = ''.join([symbol.text for symbol in word.symbols])
-                        all_text = all_text + word_text + ' '
-                        if detale:
-                            print('Word text: {} (confidence: {})'.format(word_text, word.confidence))
-
-                        for symbol in word.symbols:
-                            if detale:
-                                print('\tSymbol: {} (confidence: {})'.format(symbol.text, symbol.confidence))
-        #print(all_text)
-        '''  
-
-    if str(typ)=="obiekty_sceny":
-        # Performs label detection on the image file
-        response = client.label_detection(image=image)
-        labels = response.label_annotations
-        print('Wykryte obiekty:')
-        for label in labels:
-            print(label.description)
-
-    # response = client.
-
-    if jsonx:
-        print("")
-        print(response)
-        print("")
-        plik = open('odp.json', 'w')
-        plik.write(str(response))
-        plik.close()
-    
-    return odp
 
 def show_similar_images_from_db(start_image_idx, n_inputs=5, n_neighbors=10):
     input_images = images[start_image_idx:start_image_idx+n_inputs]
@@ -106,10 +28,6 @@ def show_similar_images_from_db(start_image_idx, n_inputs=5, n_neighbors=10):
     fig.savefig("02-image-search-grid.png")
 
 
- 
-
-
-
 def images_google(path):
     filepath=os.path.abspath(path)
     searchUrl = 'http://www.google.hr/searchbyimage/upload'
@@ -125,9 +43,11 @@ def images_google(path):
 
 CLEANR = re.compile('<.*?>') 
 
+
 def cleanhtml(raw_html):
   cleantext = re.sub(CLEANR, '', raw_html)
   return cleantext
+
 
 def text(url):
     page = requests.get(url)
@@ -167,23 +87,3 @@ def text(url):
         author = soup.find(class_='sc-1mskw74-0 sc-7eqdwf-0 ihSKrH sc-6fk6fo-0 eLtABC').text
         publication = soup.find('time').text
     return publication, title, lead, author, content, serwis, img_src
-    
-
-
-# Press the green button in the gutter to run the script.
-if __name__ == '__main__':
-    analiza_czasu=True
-    detale_analizy=True
-    pokaz_json=True
-    typ_analizy= "obiekty_sceny"
-    print("")
-    images_google('/shopping.png')
-    print("")
-
-    url="https://wydarzenia.interia.pl/raporty/raport-ukraina-rosja/aktualnosci/news-przewodow-pogrzeb-pierwszej-ofiary-wybuchu-rakiety,nId,6421383?parametr=zobacz_takze"
-    url2="https://wiadomosci.wp.pl/idzie-kryzys-dudek-nie-bedzie-oszczednosci-bez-ograniczenia-programow-spolecznych-6835377261013664a"
-    url3="https://wiadomosci.onet.pl/tylko-w-onecie/polka-odwiedzila-rosje-i-nie-ma-zludzen-zachod-widzi-to-co-chce-widziec/8vxwvxp"
-    url4="https://www.o2.pl/biznes/ludzie-przecieraja-oczy-na-jarmarku-w-gdansku-cena-grochowki-szokuje-6835661165365952a"
-    text(url4)
-    #odp=detect_document("/shopping.png",typ_analizy,detale_analizy,pokaz_json)
-# See PyCharm help at https://www.jetbrains.com/help/pycharm/
